Ajout d'un fichier Auth manager qui va m'aider à gérer la session JWT (dev en cours pour l'instant il m'aide à delete le token mais le résultat ne persiste pas comme j'ai envie) et ajout de la dépendance des CSRF token qui sera gérée via twig et php car j'avais oublié de l'ajouter, on aura donc à terme 2 token, le CSRF pour se proteger des attaques de forgery de requête inter-sites et le JWT en tant que mécanisme d'authentification et d'autorisation

+++ src/Controller/admin/UserController.php
@@ -80,6 +80,9 @@ final class UserController extends AbstractController
     #[Route('/{id}/edit', name: 'edit', requirements: ['id' => Requirement::DIGITS])]
     public function edit(User $user, Request $request, EntityManagerInterface $em): Response
     {
+        // On vérifie que l'utilisateur a bien un token valide pour accéder à la page
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
         if($form->isSubmitted() && $form->isValid()){
@@ -98,6 +101,9 @@ final class UserController extends AbstractController
     #[Route('/add', name: 'add')]
     public function add(Request $request, EntityManagerInterface $em): Response
     {
+        // On vérifie que l'utilisateur a bien un token valide pour accéder à la page
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
         $user = new User();
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
@@ -118,6 +124,9 @@ final class UserController extends AbstractController
     #[Route('/{id}/delete', name: 'delete', requirements: ['id' => '\d+'])]
     public function delete(User $user, Request $request, EntityManagerInterface $em): Response
     {
+        // On vérifie que l'utilisateur a bien un token valide pour accéder à la page
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
         $em->remove($user);
         $em->flush();
         $this->addFlash('success',"L'utilisateur a bien été supprimé");
